Added parallel module
